[train] ang

+++ TrainAndTest/train_egs_e2e_domain.py
@@ -233,8 +233,8 @@ def train(train_loader, model, ce, optimizer, epoch, scheduler):
         #
         #     loss = (1 - alpha_t) * loss + alpha_t * predict_loss + args.beta * torch.mean(-torch.log(mp))
 
-        minibatch_correct = prec * len(feats)
-        minibatch_acc = prec
+        minibatch_correct = float(prec * len(feats))
+        minibatch_acc = float(prec)
         correct += minibatch_correct
 
         total_datasize += len(feats)
